@@ -4,14 +4,16 @@
   "proxy": "http://localhost:8080",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.11.3",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@mui/material": "^7.3.6",
     "prismjs": "^1.23.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3",
     "react-file-base64": "^1.0.3",
     "react-icons": "^4.2.0",
     "react-markdown": "^5.0.3",
-    "react-redux": "^7.2.3",
+    "react-redux": "^9.2.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^5.0.1",
     "react-simple-code-editor": "^0.11.0",
