@@ -1,48 +1,165 @@
-import { Checkbox } from '@material-ui/core';
-import React from 'react';
+import { Checkbox } from "@mui/material";
+import React from "react";
 
 const Matrix = () => {
   return (
     <div className="questions_page">
-
-     <div className="page_heading" >
-      <h1 >Matrix</h1>
+      <div className="page_heading">
+        <h1>Matrix</h1>
       </div>
 
-<h1 className="Ques" ><Checkbox/>Q1.<a href="https://practice.geeksforgeeks.org/problems/spirally-traversing-a-matrix-1587115621/1" className="Ques">Spiral traversal on a Matrix</a></h1>
-<a href="/m_1" className="answer_link"> &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer</a>
+      <h1 className="Ques">
+        <Checkbox />
+        Q1.
+        <a
+          href="https://practice.geeksforgeeks.org/problems/spirally-traversing-a-matrix-1587115621/1"
+          className="Ques"
+        >
+          Spiral traversal on a Matrix
+        </a>
+      </h1>
+      <a href="/m_1" className="answer_link">
+        {" "}
+        &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer
+      </a>
 
-<h1 className="Ques" ><Checkbox/>Q2.<a href="https://leetcode.com/problems/search-a-2d-matrix/" className="Ques">Search an element in a matrix</a></h1>
-<a href="/m_2" className="answer_link"> &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer</a>
+      <h1 className="Ques">
+        <Checkbox />
+        Q2.
+        <a
+          href="https://leetcode.com/problems/search-a-2d-matrix/"
+          className="Ques"
+        >
+          Search an element in a matrix
+        </a>
+      </h1>
+      <a href="/m_2" className="answer_link">
+        {" "}
+        &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer
+      </a>
 
-<h1 className="Ques" ><Checkbox/>Q3.<a href="https://practice.geeksforgeeks.org/problems/median-in-a-row-wise-sorted-matrix1527/1#" className="Ques">Find median in a row wise sorted matrix</a></h1>
-<a href="/m_3" className="answer_link"> &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer</a>
+      <h1 className="Ques">
+        <Checkbox />
+        Q3.
+        <a
+          href="https://practice.geeksforgeeks.org/problems/median-in-a-row-wise-sorted-matrix1527/1#"
+          className="Ques"
+        >
+          Find median in a row wise sorted matrix
+        </a>
+      </h1>
+      <a href="/m_3" className="answer_link">
+        {" "}
+        &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer
+      </a>
 
-<h1 className="Ques" ><Checkbox/>Q4.<a href="https://practice.geeksforgeeks.org/problems/row-with-max-1s0023/1" className="Ques">Find row with maximum no. of 1's</a></h1>
-<a href="/m_4" className="answer_link"> &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer</a>
+      <h1 className="Ques">
+        <Checkbox />
+        Q4.
+        <a
+          href="https://practice.geeksforgeeks.org/problems/row-with-max-1s0023/1"
+          className="Ques"
+        >
+          Find row with maximum no. of 1's
+        </a>
+      </h1>
+      <a href="/m_4" className="answer_link">
+        {" "}
+        &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer
+      </a>
 
-<h1 className="Ques" ><Checkbox/>Q5.<a href="https://practice.geeksforgeeks.org/problems/sorted-matrix2333/1" className="Ques">Print elements in sorted order using row-column wise sorted matrix</a></h1>
-<a href="/m_5" className="answer_link"> &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer</a>
+      <h1 className="Ques">
+        <Checkbox />
+        Q5.
+        <a
+          href="https://practice.geeksforgeeks.org/problems/sorted-matrix2333/1"
+          className="Ques"
+        >
+          Print elements in sorted order using row-column wise sorted matrix
+        </a>
+      </h1>
+      <a href="/m_5" className="answer_link">
+        {" "}
+        &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer
+      </a>
 
-<h1 className="Ques" ><Checkbox/>Q6.<a href="https://practice.geeksforgeeks.org/problems/max-rectangle/1" className="Ques">Maximum size rectangle</a></h1>
-<a href="/m_6" className="answer_link"> &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer</a>
+      <h1 className="Ques">
+        <Checkbox />
+        Q6.
+        <a
+          href="https://practice.geeksforgeeks.org/problems/max-rectangle/1"
+          className="Ques"
+        >
+          Maximum size rectangle
+        </a>
+      </h1>
+      <a href="/m_6" className="answer_link">
+        {" "}
+        &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer
+      </a>
 
-<h1 className="Ques" ><Checkbox/>Q7.<a href="https://www.geeksforgeeks.org/find-a-specific-pair-in-matrix/" className="Ques">Find a specific pair in matrix</a></h1>
-<a href="/m_7" className="answer_link"> &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer</a>
+      <h1 className="Ques">
+        <Checkbox />
+        Q7.
+        <a
+          href="https://www.geeksforgeeks.org/find-a-specific-pair-in-matrix/"
+          className="Ques"
+        >
+          Find a specific pair in matrix
+        </a>
+      </h1>
+      <a href="/m_7" className="answer_link">
+        {" "}
+        &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer
+      </a>
 
-<h1 className="Ques" ><Checkbox/>Q8.<a href="https://www.geeksforgeeks.org/rotate-a-matrix-by-90-degree-in-clockwise-direction-without-using-any-extra-space/" className="Ques">Rotate matrix by 90 degrees</a></h1>
-<a href="/m_8" className="answer_link"> &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer</a>
+      <h1 className="Ques">
+        <Checkbox />
+        Q8.
+        <a
+          href="https://www.geeksforgeeks.org/rotate-a-matrix-by-90-degree-in-clockwise-direction-without-using-any-extra-space/"
+          className="Ques"
+        >
+          Rotate matrix by 90 degrees
+        </a>
+      </h1>
+      <a href="/m_8" className="answer_link">
+        {" "}
+        &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer
+      </a>
 
-<h1 className="Ques" ><Checkbox/>Q9.<a href="https://practice.geeksforgeeks.org/problems/kth-element-in-matrix/1" className="Ques">Kth smallest element in a row-cpumn wise sorted matrix</a></h1>
-<a href="/m_9" className="answer_link"> &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer</a>
+      <h1 className="Ques">
+        <Checkbox />
+        Q9.
+        <a
+          href="https://practice.geeksforgeeks.org/problems/kth-element-in-matrix/1"
+          className="Ques"
+        >
+          Kth smallest element in a row-cpumn wise sorted matrix
+        </a>
+      </h1>
+      <a href="/m_9" className="answer_link">
+        {" "}
+        &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer
+      </a>
 
-<h1 className="Ques" ><Checkbox/>Q10.<a href="https://www.geeksforgeeks.org/common-elements-in-all-rows-of-a-given-matrix/" className="Ques">Common elements in all rows of a given matrix</a></h1>
-<a href="/m_10" className="answer_link"> &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer</a>
+      <h1 className="Ques">
+        <Checkbox />
+        Q10.
+        <a
+          href="https://www.geeksforgeeks.org/common-elements-in-all-rows-of-a-given-matrix/"
+          className="Ques"
+        >
+          Common elements in all rows of a given matrix
+        </a>
+      </h1>
+      <a href="/m_10" className="answer_link">
+        {" "}
+        &emsp;&emsp;&emsp;&emsp;&emsp; Link to answer
+      </a>
 
-
-<br/>
-
-      </div>
+      <br />
+    </div>
   );
 };
 
